+++ packages/multichain-account-service/src/providers/BtcAccountProvider.ts
@@ -52,6 +52,7 @@ export class BtcAccountProvider extends SnapAccountProvider {
     scopes: [BtcScope.Mainnet, BtcScope.Testnet],
     bip44: {
       deriveIndex: true,
+      deriveIndexRange: true,
     },
   };
 
@@ -105,9 +106,34 @@ export class BtcAccountProvider extends SnapAccountProvider {
   ): Promise<Bip44Account<KeyringAccount>[]> {
     assertCreateAccountOptionIsSupported(options, [
       `${AccountCreationType.Bip44DeriveIndex}`,
+      `${AccountCreationType.Bip44DeriveIndexRange}`,
     ]);
 
-    const { entropySource, groupIndex } = options;
+    const { entropySource } = options;
+
+    if (options.type === AccountCreationType.Bip44DeriveIndexRange) {
+      const { range } = options;
+      return this.withSnap(async ({ keyring }) => {
+        const accounts: Bip44Account<KeyringAccount>[] = [];
+
+        for (
+          let groupIndex = range.from;
+          groupIndex <= range.to;
+          groupIndex++
+        ) {
+          const createdAccounts = await this.#createAccounts({
+            keyring,
+            entropySource,
+            groupIndex,
+          });
+          accounts.push(...createdAccounts);
+        }
+
+        return accounts;
+      });
+    }
+
+    const { groupIndex } = options;
 
     return this.withSnap(async ({ keyring }) => {
       return this.#createAccounts({ keyring, entropySource, groupIndex });